OBEX: Add pthread_join() function to properly free resources
Add pthread_join() to properly close thread and free its resources.
Without this it was observed that resources for a thread were being
leaked due to missing call to pthread_join, leading to issue in BT
on/off stress test randomly as memory was being leaked in each
iteration.

Change-Id: I3a2b1a4187cd8b2d90bb416250f73928b6b2d72b
CRs-Fixed: 1096692