Fix: DUT start SRC streaming without SNK connection after re-enable BT
If disable the Bt while snk and src is streaming. The mstreaming
device is not set to null, the next time bt is enabled, and connect
headset. It will judge the there is steaming, and start the src
streaming. The fix will set the steaming device to zero when disable
BT.
Change-Id: I59a157a60fae12a4a8a40ec8c44e15ae21199648
CRs-Fixed: 2043154