Fix: Music has background noise if codec is aptX
In the a2dp src/sink relay feature, the way that send audio data in
app layer is diffrent. In fluoride stack, the aptx run a thread to
read the incoming data, is diffrent from the way of SBC, so it cause
noise. The fix judeg the src codec config, if it'a configured as aptx,
It will follow the formal way to write audio data.

Change-Id: Idcdbde7f410082d95dd7f580b98ff5240b2974e4
CRs-Fixed: 2045155