Fix: Donnot remove the total sbc frames in header
If retrive RTP info is enabled, do not remove the total frames
numbers in headers. If retrive RTP info is disabled, when return
the raw sbc data, store the total sbc frame number in the beginning
of data. when enque the data, will remove this content in app layer.

Change-Id: I4f3cd26a768b33458e026ea74fc1707f72cb2e1e
CRs-Fixed: 2061616