OPP server: Object information is parsed at the first push indication.
In opp server, a continue is being sent to the client before
opening the object. As a result, client is pushing data before user
accepts/rejects the data. As their is data before user accepts the data,
the server forbids the data for client. This results in opp transfer failure.

Change-Id: I0ac77d2f5ed984cfa1d6d35becae19834496eba5