BTAPP: AM: Modified the audio hal library loading in audio manager.
Somes times audio hal is loading after BT ON. due to that seeing
audio issues or set params for audio is failing.
To fix this now we are loading the audio hal library, once library
loading done, then sending profile started message to GAP thread.
With this we can fix the audio issues after BT ON.

Change-Id: If2b617182464bb273606a8bf91a1fdae938c61b2