BTAPP: Device name update for SMP Only pairing devices.
For SMP pairing we don't get the Remote device name in
pairing sequence,due to that we are not saving RD name
in Remote device properties and not showing name in
bonded list.
To fix it we should save the remote device name from
inquiry DB in Remote device properties, if RD properties
are null.

Change-Id: Id9884d481101fb850316698ff9b4a07fcb5cd4c5