Gatt Server: Send connect request to client on transport 2
When ever a client is connected and nottification reaches app we
send a connect request to client. When a connect request is sent to
a device which already has a BR/EDR connection established with transport
as auto connect. A new gatt connection is getting established over
BR/EDR with the same device resulting in 2 connection id with same device.

Change-Id: I80c53d613091ce958142948b694e9420a5c06ee3