BTAPP: Handle the HOGP device pairing cancel in app
For HOGP devices, we just need to show passkey
and auto accept the SSP. However we are showing yes/no
option and not canceling the bonding procedure if user
enters no option.
To fix it in BT APP, we should show passkey and cancel
option only for passkey notification devices like
HOGP devices and send ssp reply with auto accept.
Also for cancel option, need to send cancel bond to
cancel bonding procedure.

Change-Id: I430876204035bcc9990ae1eec6d5527837f150ee