BT APP: Format to display 6 digit SSP key to user
while generating the SSP key few times it is possible
that passkey value is less than 6 digit, due to that
while bonding with hogp display bonding fails.
To fix it even though we get less than 6 digit passkey
to application we should format and show 6 digit with
adding 0 as trailing to passkey.

Change-Id: I271b5c95bcb7cd9da7d2d83f9413be0aeb7d4a01